requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==23.2.0
beautifulsoup4==4.12.3
binance-connector==3.8.1
binance-futures-connector==4.0.0
blinker==1.8.2
ccxt==4.3.70
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.8
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.29.0
setuptools==72.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.45.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.2.2
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
